+++ example/gauge/gauge.dart
@@ -4,7 +4,7 @@
 import 'dart:html';
 import 'dart:async';
 import 'dart:math';
-import 'package:google_visualization_api/google_visualization_api.dart'
+import 'package:google_charts/google_charts.dart'
     show Gauge, DataTable, arrayToDataTable;
 
 void main() {