+++ example/polymer/gviz_pie_chart.dart
@@ -1,4 +1,4 @@
-import 'package:google_visualization_api/google_visualization_api.dart'
+import 'package:google_charts/google_charts.dart'
     show PieChart,  arrayToDataTable;
 import 'package:polymer/polymer.dart';