+++ lib/google_charts.dart
@@ -1,0 +1,11 @@
+library google.charts;
+
+export 'static.dart';
+export 'events.dart';
+export 'base.dart';
+// Classic Visualization API
+export 'visualization.dart';
+// Material Bar Charts
+export 'charts.dart' hide Chart;
+
+//export 'gen/chart_options.dart';