+++ lib/google_visualization_api.dart
@@ -1,7 +1,0 @@
-library google_visualization_api;
-
-export 'static.dart';
-export 'events.dart';
-export 'base.dart';
-export 'charts.dart';
-//export 'gen/chart_options.dart';