+++ lib/src/visualization/scatter_chart.dart
@@ -9,14 +9,3 @@ class ScatterChart extends Chart with _DiffChart {
     return Chart.load(packages: ['corechart']);
   }
 }
-
-/// Experimental
-class Scatter extends Chart {
-  Scatter(Element e) : super._(e, "Scatter", charts) {
-
-  }
-
-  static Future load() {
-    return Chart.load(packages: ['scatter'], version: "1.1");
-  }
-}